+++ Assets/Scripts/Main/Enemies.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 12edf4a1dd09e4c599b609b72da0a8f5
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
